backend/pyproject.toml
[tool.poetry]
name = "reviflow-backend"
version = "0.1.0"
description = "Backend for Reviflow EdTech App"
authors = ["Michael <michael@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
pydantic = "^2.6.0"
pydantic-settings = "^2.1.0"
sqlmodel = "^0.0.14"
fastapi-users = "^13.0.0"
fastapi-users-db-sqlalchemy = "^6.0.0"
aiosqlite = "^0.19.0"
asyncpg = "^0.29.0"
python-multipart = "^0.0.9"
openai = "^1.10.0" # For OpenRouter
httpx = "^0.26.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
